Move timeline to own package
